requirements.txt
fastapi==0.95.1
jose==1.0.0
openai==0.27.2
passlib==1.7.4
pydantic==1.10.7
python-dotenv==1.0.0
python_jose==3.3.0
SQLAlchemy==1.4.46
tenacity==8.2.2
Utils==1.0.1
uvicorn==0.21.1
